test cleanup and more reorg of multiplier code
